refactor(llm): remove LiteLLM sidecar module and restore 100% coverage (#2075)
Delete litellm.py (LiteLLMWatchdog, LiteLLMRenderer) and test_litellm.py.
Move parse_model_entry into llm_router.py. Remove all litellm_watchdog
references from main.py, test_main.py, conftest, and _e2e_env.
Restore --cov-fail-under=100.
