+++ src/klangk/pyproject.toml
@@ -51,7 +51,7 @@ klangk = "klangk.cli.main:main"
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 asyncio_default_fixture_loop_scope = "function"
-addopts = "--capture=no --cov=klangk --cov-report=term-missing --no-cov-on-fail --cov-fail-under=99 --timeout=60"
+addopts = "--capture=no --cov=klangk --cov-report=term-missing --no-cov-on-fail --cov-fail-under=100 --timeout=60"
 # Treat new warnings as errors so they can't re-accumulate (#1493). The
 # two ignores are pre-existing mock-induced GC artifacts (MagicMock stubs
 # whose coroutine objects are collected in a later test) -- not real bugs
